+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All important changes to this project will be documented in this file.
 * Added handling `.cmd` script and shortcut to `.bat` or `.cmd` script as start file on Windows
 * Added editing messages from Minecraft in Discord channel and edit webhook messages from Discord ([#27](https://github.com/Druzai/Bot_Mc_discord/issues/27))
 * Added check for edited messages whether they are stored in cache or not
+* Added timeout to stop server when no players found during long period of time ([#29](https://github.com/Druzai/Bot_Mc_discord/issues/29))
 
 #### Fixed:
 
@@ -20,6 +21,7 @@ All important changes to this project will be documented in this file.
 * Fixed slow bot response when sending a message from synchronous function
 * Added additional spaces when there are more than 9 items in list to display in command output
 * Reduced `backup force` command cooldown to 15 seconds
+* Fixed and rewrote many translations
 
 #### Breaking changes:
 
@@ -30,6 +32,7 @@ All important changes to this project will be documented in this file.
     * From `servers show` to `server`
     * From `ops history` to `op history`
     * From `ops` to `op info`
+    * From `forceload` to `schedule forceload`
   * Created commands: `associate add` and `associate del`
   * Added alias `associate` to `assoc`
   * Added alias `ls` to all `list` subcommands